move prng test to others
